refactor: split out wordListParser from compiler
