+++ packages/cspell-tools/src/compiler/CompileOptions.ts
@@ -1,27 +1,7 @@
-export interface CompileCommonOptions {
-    output?: string;
-    compress: boolean;
-    /**
-     * @deprecated Use maxDepth
-     */
-    max_depth?: string;
-    maxDepth?: string;
-    merge?: string;
-    experimental: string[];
-    split?: boolean;
-    sort?: boolean;
-    keepRawCase?: boolean;
-    trie?: boolean;
-    trie3?: boolean;
-    trie4?: boolean;
-    trieBase?: string;
-    useLegacySplitter?: boolean;
-}
-export interface CompileOptions extends CompileCommonOptions {
-    sort: boolean;
+export interface CompileOptions {
+    skipNormalization: boolean | undefined;
+    splitWords: boolean | undefined;
     keepRawCase: boolean;
-}
-export interface CompileTrieOptions extends CompileCommonOptions {
-    trie3: boolean;
-    trie4: boolean;
+    sort: boolean;
+    legacy: boolean | undefined;
 }
